src/iot-consumer/requirements.txt
pika
boto3
python-dotenv